refactor: move `InPrSuppression` types from `recap.ts` to `types.ts`
Addresses manki-review review comment on PR #584
https://github.com/manki-review/manki/pull/584#discussion_r3105196299
