+++ src/recap.ts
@@ -3,7 +3,7 @@ import * as github from '@actions/github';
 import { ClaudeClient } from './claude';
 import { titleToSlug } from './github';
 import { matchesSuppression, Suppression } from './memory';
-import { AuthorReplyClass, Finding, FindingFingerprint, FindingSeverity, migrateLegacySeverity, SEVERITY_TOKEN_PATTERN } from './types';
+import { AuthorReplyClass, Finding, FindingFingerprint, FindingSeverity, InPrSuppression, InPrSuppressionReason, migrateLegacySeverity, SEVERITY_TOKEN_PATTERN } from './types';
 
 type Octokit = ReturnType<typeof github.getOctokit>;
 
@@ -108,15 +108,6 @@ interface PreviousFinding {
   threadId?: string;
   threadUrl?: string;
   authorReplyText?: string;
-}
-
-/** Reason why a prior-round thread's fingerprint is suppressing current findings. */
-export type InPrSuppressionReason = 'resolved-thread' | 'agree-reply';
-
-/** Fingerprint-level suppression derived from the current PR's thread state. */
-export interface InPrSuppression {
-  fingerprint: FindingFingerprint;
-  reason: InPrSuppressionReason;
 }
 
 /**
